Extract `CodeBlock` component into own file
