+++ website/src/components/landing.js
@@ -11,7 +11,7 @@ import overlayLegacy from '../images/pattern_landing_legacy.png'
 import Grid from './grid'
 import { Content } from './main'
 import Button from './button'
-import CodeBlock from './code'
+import CodeBlock from './codeBlock'
 import { H1, H2, H3 } from './typography'
 import Link from './link'
 import classes from '../styles/landing.module.sass'
